// frontend/lib/generated-api/models/ProjectDetailed.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI Analyst API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Project } from './Project';
import { ProjectFromJSON, ProjectFromJSONTyped, ProjectToJSON, ProjectToJSONTyped } from './Project';
import type { WorkflowRunDetailed } from './WorkflowRunDetailed';
import {
  WorkflowRunDetailedFromJSON,
  WorkflowRunDetailedFromJSONTyped,
  WorkflowRunDetailedToJSON,
  WorkflowRunDetailedToJSONTyped,
} from './WorkflowRunDetailed';

/**
 *
 * @export
 * @interface ProjectDetailed
 */
export interface ProjectDetailed {
  /**
   *
   * @type {Project}
   * @memberof ProjectDetailed
   */
  project: Project;
  /**
   *
   * @type {WorkflowRunDetailed}
   * @memberof ProjectDetailed
   */
  workflowRun?: WorkflowRunDetailed | null;
}

/**
 * Check if a given object implements the ProjectDetailed interface.
 */
export function instanceOfProjectDetailed(value: object): value is ProjectDetailed {
  if (!('project' in value) || value['project'] === undefined) return false;
  return true;
}

export function ProjectDetailedFromJSON(json: any): ProjectDetailed {
  return ProjectDetailedFromJSONTyped(json, false);
}

export function ProjectDetailedFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProjectDetailed {
  if (json == null) {
    return json;
  }
  return {
    project: ProjectFromJSON(json['project']),
    workflowRun: json['workflow_run'] == null ? undefined : WorkflowRunDetailedFromJSON(json['workflow_run']),
  };
}

export function ProjectDetailedToJSON(json: any): ProjectDetailed {
  return ProjectDetailedToJSONTyped(json, false);
}

export function ProjectDetailedToJSONTyped(value?: ProjectDetailed | null, ignoreDiscriminator: boolean = false): any {
  if (value == null) {
    return value;
  }

  return {
    project: ProjectToJSON(value['project']),
    workflow_run: WorkflowRunDetailedToJSON(value['workflowRun']),
  };
}